requirements.txt
fastapi==0.104.1
uvicorn[standard]
pydantic==2.5.0
python-dotenv==1.0.0
openai>=2.8.0
anthropic>=0.39.0
google-generativeai>=0.8.0
python-multipart==0.0.6
Pillow>=10.0.0
reportlab>=4.0.0
python-docx>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0